+++ src/Support/EnsureValidPassphrase.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace CodebarAg\DocuWare\Support;
+
+use CodebarAg\DocuWare\Exceptions\UnableToFindPassphrase;
+
+class EnsureValidPassphrase
+{
+    public static function check(): void
+    {
+        throw_if(
+            empty(config('docuware.passphrase')),
+            UnableToFindPassphrase::create(),
+        );
+    }
+}